Cast auxiliary rays perpendicularly by distance
Previously using an angle to "rotate" the auxiliary rays cast after
"seeing" a corner caused artifacts when the angle between the initial
ray and a wall past the corner was very small, resulting in the
auxiliary ray striking that wall instead of "peeking" past the corner as
expect.

By stepping away from the corner, perpendicularly to the ray, but a
fixed step distance, we eliminate this artifact and consistently "see"
past corners as expected.