Toss out optimization, build fov with lots of rays
Instead of optimizing the view cone by only casting rays at corners, we
now simply throw out a bunch of rays at regular increments throughout
the view cone and collect all of those points. While the result is a lot
more rays, we no longer have the problem of trying to find arc-ray
intersections to complete the cone, nor do we have to do any "funny
business" to give the cone a curve.

The code is now a lot simpler, and we can always come back IFF we need
to optimize this